HUGE wip to implement NRT (nullable reference type) on every project and solve every warning, WIP

+++ src/MithrilShards.Chain.Bitcoin/Network/NodeServices.cs
@@ -13,7 +13,7 @@ namespace MithrilShards.Chain.Bitcoin.Network
    // BIP process.
    /// </summary>
    [Flags]
-   public enum NodeServices : ulong
+   public enum NodeServices : uint
    {
       /// <summary>
       /// NODE_NETWORK means that the node is capable of serving the block chain. It is currently
